CHP. 1-LAB 1
For this lab I was learning how to use the paragraph, head, body, hr, and title element to create a template outline for a website ideal. The goal within this design was to help propose a plan for striking a cord. I had to add a paragraph for my body, and added my name as the creator for the site.